pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
authlib = "^1.3.0"
requests = "^2.31.0"
redis = "^5.0.1"
django = "^4.2.5"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
celery = "^5.3.6"
pytz = "^2024.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
aiohttp = "^3.9.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
